用线性加速度计实现了捕捉功能，ver1.0完成！

+++ app/src/main/java/com/example/mydatabinding/model/IDBModel.java
@@ -13,7 +13,7 @@ public interface IDBModel {
     void insertTrainer(String t_name,String password);//用户注册
     //暂不提供删除方法
     void selectPokemonByTrainer(long trainer_id);//根据用户ID查询其拥有的宝可梦ID
-    void insertPokemonByTrainer(long trainer_id,long pokemon_id);//增加用户捕捉到的一只宝可梦
+    void insertPokemonByTrainer(long trainer_id,int pokemon_id);//增加用户捕捉到的一只宝可梦
 
     void isLogin(Context context);//查询本地SQLite数据库
     void saveLogin(Trainer trainer, Context context);//保存登录状态
